+++ src/generate/device.rs
@@ -35,14 +35,6 @@ pub fn render(
         });
     }
 
-    if target != Target::None && target != Target::CortexM && target != Target::RISCV {
-        out.extend(quote! {
-            #![cfg_attr(feature = "rt", feature(global_asm))]
-            #![cfg_attr(feature = "rt", feature(use_extern_macros))]
-            #![cfg_attr(feature = "rt", feature(used))]
-        });
-    }
-
     out.extend(quote! {
         #![doc = #doc]
         // Deny a subset of warnings
@@ -84,8 +76,6 @@ pub fn render(
                 extern crate msp430;
                 #[cfg(feature = "rt")]
                 extern crate msp430_rt;
-                #[cfg(feature = "rt")]
-                pub use msp430_rt::default_handler;
             });
         }
         Target::RISCV => {
@@ -156,6 +146,17 @@ pub fn render(
                 };
             });
         }
+    }
+
+    if target == Target::Msp430 {
+        out.extend(quote! {
+            // XXX: Are there any core peripherals, really? Requires bump of msp430 crate.
+            // pub use msp430::peripheral::Peripherals as CorePeripherals;
+            #[cfg(feature = "rt")]
+            pub use msp430_rt::interrupt;
+            #[cfg(feature = "rt")]
+            pub use self::Interrupt as interrupt;
+        });
     }
 
     let generic_file = std::str::from_utf8(include_bytes!("generic.rs")).unwrap();