+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- [breaking-change] Among other cleanups, MSP430 crates are now expected to
+  use the `msp430_rt::interrupt` attribute macro and `device.x` for interrupt
+  support. The `INTERRUPT` array has been renamed `__INTERRUPT`.
+
 ## [v0.17.0] - 2019-12-31
 
 ### Fixed